+++ capa/features/extractors/dnfile/file.py
@@ -13,17 +13,21 @@ from typing import TYPE_CHECKING, Tuple, Iterator
 if TYPE_CHECKING:
     import dnfile
     from capa.features.common import Feature, Format, String
-    from capa.features.file import Import
+    from capa.features.file import Import, FunctionName
 
 import capa.features.extractors
 
 
 def extract_file_import_names(pe: dnfile.dnPE) -> Iterator[Tuple[Import, int]]:
-    yield from capa.features.extractors.dotnetfile.extract_file_import_names(pe)
+    yield from capa.features.extractors.dotnetfile.extract_file_import_names(pe=pe)
 
 
 def extract_file_format(pe: dnfile.dnPE) -> Iterator[Tuple[Format, int]]:
     yield from capa.features.extractors.dotnetfile.extract_file_format(pe=pe)
+
+
+def extract_file_function_names(pe: dnfile.dnPE) -> Iterator[Tuple[FunctionName, int]]:
+    yield from capa.features.extractors.dotnetfile.extract_file_function_names(pe=pe)
 
 
 def extract_file_strings(pe: dnfile.dnPE) -> Iterator[Tuple[String, int]]:
@@ -38,7 +42,7 @@ def extract_features(pe: dnfile.dnPE) -> Iterator[Tuple[Feature, int]]:
 
 FILE_HANDLERS = (
     extract_file_import_names,
+    extract_file_function_names,
     extract_file_strings,
-    # TODO extract_file_function_names,
     extract_file_format,
 )