+++ capa/features/extractors/dotnetfile.py
@@ -1,12 +1,11 @@
 import logging
 from typing import Tuple, Iterator
-from itertools import chain
 
 import dnfile
 import pefile
 
 import capa.features.extractors.helpers
-from capa.features.file import Import
+from capa.features.file import Import, FunctionName
 from capa.features.common import (
     OS,
     OS_ANY,
@@ -20,7 +19,12 @@ from capa.features.common import (
     Feature,
 )
 from capa.features.extractors.base_extractor import FeatureExtractor
-from capa.features.extractors.dnfile.helpers import get_dotnet_managed_imports, get_dotnet_unmanaged_imports
+from capa.features.extractors.dnfile.helpers import (
+    get_dotnet_managed_imports,
+    calculate_dotnet_token_value,
+    get_dotnet_unmanaged_imports,
+    get_dotnet_managed_method_names,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -30,15 +34,20 @@ def extract_file_format(**kwargs) -> Iterator[Tuple[Format, int]]:
 
 
 def extract_file_import_names(pe: dnfile.dnPE, **kwargs) -> Iterator[Tuple[Import, int]]:
-    for (token, imp) in chain(get_dotnet_managed_imports(pe), get_dotnet_unmanaged_imports(pe)):
-        if "::" in imp:
-            # like System.IO.File::OpenRead
-            yield Import(imp), token
-        else:
-            # like kernel32.CreateFileA
-            dll, _, symbol = imp.rpartition(".")
-            for symbol_variant in capa.features.extractors.helpers.generate_symbols(dll, symbol):
-                yield Import(symbol_variant), token
+    for (token, name) in get_dotnet_managed_imports(pe):
+        # like System.IO.File::OpenRead
+        yield Import(name), token
+
+    for (token, name) in get_dotnet_unmanaged_imports(pe):
+        # like kernel32.CreateFileA
+        dll, _, symbol = name.rpartition(".")
+        for name_variant in capa.features.extractors.helpers.generate_symbols(dll, symbol):
+            yield Import(name_variant), token
+
+
+def extract_file_function_names(pe: dnfile.dnPE, **kwargs) -> Iterator[Tuple[FunctionName, int]]:
+    for (token, name) in get_dotnet_managed_method_names(pe):
+        yield FunctionName(name), token
 
 
 def extract_file_os(**kwargs) -> Iterator[Tuple[OS, int]]:
@@ -68,8 +77,8 @@ def extract_file_features(pe: dnfile.dnPE) -> Iterator[Tuple[Feature, int]]:
 
 FILE_HANDLERS = (
     extract_file_import_names,
+    extract_file_function_names,
     extract_file_strings,
-    # TODO extract_file_function_names,
     extract_file_format,
 )